backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.7.1
PyJWT==2.8.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
bcrypt==4.1.3
passlib[bcrypt]==1.7.4

# AI — Groq (OpenAI-compatible; free tier at console.groq.com)
httpx>=0.25.0,<0.28.0
openai==1.35.3

# Payments
stripe==10.1.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
